requirements.txt
absl-py==1.4.0
alembic==1.11.1
appdirs==1.4.4
astunparse==1.6.3
attrs==23.1.0
autoray==0.6.6
bayesian-optimization==1.4.3
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.7.0
braceexpand==0.1.7
bs4==0.0.1
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chex==0.1.7
click==8.1.3
clingo==5.6.2
cloudpickle==2.2.1
cma==3.3.0
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
contourpy==1.0.7
cotengra==0.5.0
cpmpy==0.9.15
cryptography==41.0.3
cycler==0.11.0
Cython==0.29.34
cytoolz==0.12.2
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.6
dm-clrs==1.0.0
dm-haiku==0.0.9
dm-tree==0.1.8
docker-pycreds==0.4.0
easydict==1.10
equinox==0.10.11
etils==1.3.0
evojax==0.2.15
evosax==0.1.4
fabric==3.2.1
filelock==3.12.0
flatbuffers==23.5.9
flax==0.6.10
fonttools==4.39.3
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
graphviz==0.20.1
grpcio==1.54.2
gurobipy==10.0.2
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
invoke==2.2.0
jax==0.4.14
jaxlib==0.4.14
jaxtyping==0.2.21
Jinja2==3.1.2
jmp==0.0.4
joblib==1.2.0
kahypar==1.3.4
keras==2.13.1rc0
kiwisolver==1.4.4
libclang==16.0.0
llvmlite==0.41.0
LTNtorch==1.0.1
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
maturin==0.15.2
mdurl==0.1.2
minizinc==0.9.0
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.5
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nevergrad==0.13.0
numba==0.58.0
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.5
optuna==3.2.0
orbax-checkpoint==0.2.4
ortools==9.6.2534
packaging==23.1
pandas==2.0.1
paramiko==3.3.1
pathspec==0.11.1
pathtools==0.1.2
Pillow==9.5.0
platformdirs==3.9.1
plotly==5.14.1
pluggy==1.3.0
problog==2.2.4
promise==2.3
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyeda==0.28.0
Pygments==2.15.1
pyinstrument==4.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
PySDD==0.2.11
pytest==7.4.1
python-dateutil==2.8.2
python-sat==0.1.8.dev7
pytz==2023.3
PyYAML==6.0
quimb==1.6.0
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.20.0
setproctitle==1.3.2
six==1.16.0
sklearn==0.0.post4
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow==2.13.0rc0
tensorflow-estimator==2.13.0rc0
tensorflow-macos==2.13.0rc0
tensorflow-metadata==1.13.1
tensorflow-probability==0.20.0
tensorly==0.8.1
tensornetwork==0.4.6
tensorstore==0.1.37
termcolor==2.3.0
tfds-nightly==4.5.2.dev202204190046
threadpoolctl==3.1.0
tntorch==1.1.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
transformers==4.30.2
typeguard==4.1.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.15
wandb==0.15.5
webdataset==0.2.48
Werkzeug==2.3.4
wget==3.2
wrapt==1.14.1